+++ src/field/trigger.rs
@@ -14,7 +14,7 @@ pub struct Trigger {
 
 pub fn ptr_table(rom_data: &[u8]) -> Result<ptr_table::Table, Box<Error>> {
     // TODO bounds check rom data.
-    let addr = rom_map::snes_to_file(rom_map::EVENT_TRIGGER_POINTERS) as usize;
+    let addr = rom_map::snes_to_file(rom_map::EVENT_TRIGGER_POINTERS);
     let table = ptr_table::Table::new(
         &rom_data[addr..],
         0x1a0,
